Add extra echos before the DB drop command
In case someone like @suecarmol runs the command in production by accident

+++ bin/restore.sh
@@ -9,7 +9,16 @@ restore_file=${1}
 if /app/bin/django_wait_for_db.sh
 then
 
-    echo "Dropping existing DB."
+    echo "This will drop the DB. Proceed [y/N]?"
+    read -p "This will drop the DB. Proceed [y/N]?" -n 1 -r
+    echo ""
+    if [[ ! $REPLY =~ ^[Yy]$ ]]
+    then
+        echo "Exiting..."
+        exit
+    fi
+
+    echo "Dropping existing DB"
     mysql -h db -u root -p"${MYSQL_ROOT_PASSWORD}" -D "${MYSQL_DATABASE}" -e "DROP DATABASE ${MYSQL_DATABASE}; CREATE DATABASE ${MYSQL_DATABASE};" | :
 
     echo "Importing backup DB."
